Split post-merge audit summary modules
